+++ Line/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2024.2.14.1
+* Remove deprecated callbacks `onRewardedAdVideoStarted()` and `onRewardedAdVideoCompleted()`.
+* Remove passing of `Activity` contexts into LINE APIs.
+
 ## 2024.2.14.0
 * Certified with Line SDK 2.7.20240214.